backend/requirements.txt
# Core dependencies
ollama>=0.6.1
openai>=1.0.0
google-genai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector>=0.2.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
tenacity>=8.2.0
python-json-logger>=2.0.0
tqdm>=4.66.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.19

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

# Jupyter (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
